Split into request table and result table for CATV requests
Because the results are stored as JSONB in the table, and the result size is significantly high, ~20-30 MB, storing results and requests in the same table would make querying for the request list significantly slower.
Possible further improvement: create a FDW for the result table and store it on S3. Need to upgrade to Postgres 11.4 and install aws_s3 extension on our database